Use customer_id instead of _sdc_customer_id for performance reports and cast ids to bigint instead of int

+++ models/base/adwords_ad_groups.sql
@@ -9,21 +9,21 @@ renamed as (
     select
 
         -- PK
-        id                     as ad_group_id,
+        id                        as ad_group_id,
 
         -- FKeys
-        _sdc_customer_id::int  as account_id,
-        campaign_id            as campaign_id,
+        _sdc_customer_id::bigint  as account_id,
+        campaign_id               as campaign_id,
 
-        base_ad_group_id       as base_ad_group_id,
-        base_campaign_id       as base_campaign_id,
+        base_ad_group_id          as base_ad_group_id,
+        base_campaign_id          as base_campaign_id,
 
         -- Ad Group Info
-        campaign_name          as campaign_name,
-        name                   as ad_group_name,
+        campaign_name             as campaign_name,
+        name                      as ad_group_name,
 
-        ad_group_type          as ad_group_type,
-        status                 as ad_group_status
+        ad_group_type             as ad_group_type,
+        status                    as ad_group_status
 
     from source
